+++ src/components/forms/fields/FormCheckbox.tsx
@@ -8,13 +8,13 @@ export interface FormCheckboxProps extends Omit<React.InputHTMLAttributes<HTMLIn
 }
 
 export const FormCheckbox = (props: FormCheckboxProps) => {
-  const { value, onChange, isInvalid, error, label, required, mergedProps, formId } = useFormField(props);
+  const { value, onChange, isInvalid, error, label, required, mergedProps, formId, className } = useFormField(props);
 
   const errorId = isInvalid && error ? `${formId}-${props.name}-error` : undefined;
   const controlId = `${formId}-${props.name}`;
 
   return (
-    <Form.Group controlId={controlId}>
+    <Form.Group controlId={controlId} className={`mt-2${className ? ` ${className}` : ''}`}>
       {isInvalid && error && (
         <Form.Text id={errorId} className="text-danger">
           {error}
@@ -41,13 +41,13 @@ export const FormCheckbox = (props: FormCheckboxProps) => {
 };
 
 export const FormSwitch = (props: FormCheckboxProps) => {
-  const { value, onChange, isInvalid, error, label, required, mergedProps, formId } = useFormField(props);
+  const { value, onChange, isInvalid, error, label, required, mergedProps, formId, className } = useFormField(props);
 
   const errorId = isInvalid && error ? `${formId}-${props.name}-error` : undefined;
   const controlId = `${formId}-${props.name}`;
 
   return (
-    <Form.Group controlId={controlId}>
+    <Form.Group controlId={controlId} className={className}>
       {isInvalid && error && (
         <Form.Text id={errorId} className="text-danger">
           {error}