+++ change-log.md
@@ -347,3 +347,7 @@
 - Make `onChange` argument of `FormInput` optional as it can get it from the form context
 - No longer show "Select" option in number of rows dropdown of `DataTables` component
 - New `formId` in form context that is used to generate unique `controlIds` in case of multiple forms
+
+##### Version 0.9.6
+- All form input components accept `className` prop that is given to `Form.Group` component
+- Add 'mt-2' class to `FormCheckbox` improve vertical spacing